% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{calc_hydroGOF}
\alias{calc_hydroGOF}
\title{Wrapper function around \code{\link{hydroGOF}} functions}
\usage{
calc_hydroGOF(GOF_fun, Qsim, Qobs, na.rm = TRUE)
}
\arguments{
\item{GOF_fun}{a function, or (todo consider only a functional) a string with
function name, of the format \code{GOF_fun(Qsim, Qobs, na.rm = "TRUE")},
typically from the \code{\link{hydroGOF}} package}

\item{Qsim}{vector, matrix, data.frame etc of simulated runoff values}

\item{Qobs}{vector, matrix, data.frame etc of observed runoff values}

\item{na.rm}{a logical value indicating if NA should be removed}
}
\value{
transformed runoff in same format as input
}
\description{
Calculates Goodness-of-Fit functions for two runoff series
}
\examples{
# for the first example the function KGE must be loaded,
#  for example with library(hydroGOF)
calc_hydroGOF(KGE, 1:10, seq(0, 9))
# this is NA
calc_hydroGOF("KGE", 1:10, rep(0, 10))
# this is also NA
calc_hydroGOF(KGE, 1:10, as.numeric(rep(NA, 10)))
}
\seealso{
\code{\link{hydroGOF}}
}
