% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{calibrate_model}
\alias{calibrate_model}
\title{Model calibration}
\usage{
calibrate_model(
  hydro_data,
  split_indices,
  model,
  input,
  error_crit_transfo = "KGE_none",
  cal_maximize = TRUE,
  cal_fn = "DEoptim",
  do_transfo_param = FALSE,
  cal_par = default_cal_par
)
}
\arguments{
\item{split_indices}{a list of indices from \code{\link{split_data_set}}
containing the elements \code{ind_cal} and \code{ind_warm}}

\item{model}{a string specifying the hydrological model, currently
implemented are airGR and TUWmodel package models}

\item{input}{the output of the \code{create_input} function, dependent on the
model choice, in general containing information about date, precipitation
air temperature and potential evapotranspiration and spatial information as
area of the catchment or hypsometric curves}

\item{error_crit_transfo}{string specifying the error criterion for
calibration and the runoff transformation separated by a \code{"_"}
supported are validation criteria from the \code{\link{hydroGOF}} package
usable by the \code{\link{calc_hydroGOF}} function, for supported runoff
transformations please refer to \code{\link{transfo_q}}}

\item{cal_maximize}{a logical indicating if the calibration
error criterion should be maximized (or minimized)}

\item{cal_fn}{a string specifying optimization algorithm to be used
for the model calibration Currently \code{\link[airGR]{Calibration_Michel}}
is implemented for \code{\link{airGR}} models, and \code{\link{DEoptim}} ,
\code{\link{hydroPSO}} and \code{\link{malschains}} are supported for all
models.}

\item{do_transfo_param}{logical indicating if parameter transformation to a
hypercube should be applied during calibration}

\item{cal_par}{a list of calibration settings dependent on the calibration
function. Package default values \code{default_cal_par} can be overwritten
by users.}

\item{BasinObs}{data frame with time series of input data from
\code{\link{load_meteo_data}}}
}
\value{
a list of calibration results with elements called
\itemize{
\item \code{model_param} a vector of calibrated model parameters,
\item \code{error_crit_transfo} the error criterion and transformation used,
\item \code{error_crit_val} the value of the error criterion,
\item \code{more info} calibration function and model specific output
information.
}
}
\description{
Calibrates a hydrological model
}
\note{
Calibration_Michel does not call \code{\link{optim_fn}}, this is why
the large list of otherwise global variables is still required as input
todo: 1) check if some input is really needed for Calibration_Michel as
e.g. BasinObs
2) think about how to easily access and change cal_par values for
end user
}
\examples{
calibration_results <- calibrate_model(
  hydro_data, split_indices, model, input, error_crit_transfo, cal_maximize,
  cal_fn, do_transfo_param, cal_par
)
}
\seealso{
\code{\link{call_cal_fn}}, \code{\link{optim_fn}}
}
