% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{call_cal_fn}
\alias{call_cal_fn}
\title{Call calibration function}
\usage{
call_cal_fn(
  cal_fn,
  hydro_data,
  split_indices,
  model,
  input,
  error_crit,
  cal_maximize,
  cal_q_transfo,
  do_transfo_param,
  cal_par = default_cal_par
)
}
\arguments{
\item{cal_fn}{a string specifying optimization algorithm to be used for the
model calibration Currently \code{"DEoptim"}, \code{"hydroPSO"} and
\code{"malschains"} are supported.}

\item{hydro_data}{a list or data frame containing observed runoff loaded
with \code{\link{load_meteo_data}}}

\item{split_indices}{indices indicating the
warm-up and calibration period, i.e. the output of
\code{\link{split_data_set}}}

\item{model}{a string indicating the hydrological model}

\item{input}{model input returned by \code{\link{create_input}}}

\item{error_crit}{a string of a function indicating an error criterion, e.g.
\code{\link[hydroGOF]{KGE}}, or others from the \code{\link{hydroGOF}}
which are usable by \code{\link{calc_hydroGOF}}
functions}

\item{cal_maximize}{a logical indicating if function should be
maximized}

\item{cal_q_transfo}{a string indicating how runoff should be transformed
with \code{\link{transfo_q}}}

\item{do_transfo_param}{logical indicating
whether calibration is done with real or transformed parameters}

\item{cal_par}{a list of calibration settings dependent on the calibration
function. Package default values \code{default_cal_par} can be overwritten
by users.}
}
\description{
Call several optimization functions for model calibration which are available
as R packages
}
\examples{
cal_output <- call_cal_fn(
  cal_fn, hydro_data, split_indices, model, input,
  error_crit, cal_maximize, cal_q_transfo, do_transfo_param, cal_par
)
}
