% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\docType{data}
\name{default_cal_par}
\alias{default_cal_par}
\title{Default calibration parameters}
\format{
An object of class \code{list} of length 17.
}
\usage{
default_cal_par
}
\description{
Default calibration parameters
}
\keyword{datasets}
