% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{find_monthly_indices}
\alias{find_monthly_indices}
\title{Find Monthly Indices in a Date Vector}
\usage{
find_monthly_indices(date, months, ind = seq_along(date))
}
\arguments{
\item{date}{A vector of dates (e.g., \code{Date}, \code{POSIXt}).}

\item{months}{A character vector of two-digit month strings (e.g., \code{c("02", "03")}).}

\item{ind}{An optional vector of indices used to subset the input date vector before filtering.
Defaults to all indices in the date vector.}
}
\value{
An integer vector of indices corresponding to
}
\description{
Returns the indices of dates that fall within specified months. Useful for
subsetting time series data by month.
}
