% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{get_family}
\alias{get_family}
\title{Get the family of a model}
\usage{
get_family(model_str)
}
\arguments{
\item{model_str}{a string indicating a hydrological model}
}
\value{
the package or family name
}
\description{
Reverse lookup table to get the package/family of a hydrological model
}
\examples{
get_family("GR4J")
}
