% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\docType{data}
\name{hydro_family}
\alias{hydro_family}
\title{Hydrofamily or Package}
\format{
An object of class \code{data.frame} with 7 rows and 2 columns.
}
\usage{
hydro_family
}
\description{
Hydrofamily or Package
}
\keyword{datasets}
