% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{KGEtang}
\alias{KGEtang}
\title{Kling-Gupta Efficiency (KGE'') after Tang et al. (2021)}
\usage{
KGEtang(Qsim, Qobs, single_output = TRUE, ...)
}
\arguments{
\item{Qsim}{A numeric vector of simulated runoff values.}

\item{Qobs}{A numeric vector of observed runoff values.}

\item{single_output}{A boolean if a single combined or
all components should be output}
}
\description{
Computes the modified Kling-Gupta Efficiency (KGE) as proposed by Tang et al. (2021),
https://doi.org/10.1175/jcli-d-21-0067.1, Equation (5)
}
