% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{load_meteo_data}
\alias{load_meteo_data}
\title{Load meteo data}
\usage{
load_meteo_data(file, tzone = "UTC")
}
\arguments{
\item{file}{file}

\item{tzone}{transfers the \code{DatesR} column to POSiXct to time zone
\code{tzone} if \code{"UTC"} (default) is chosen, a Date vector is not
transferred to another time zone}
}
\value{
a data frame BasinObs
}
\description{
Loads a data frame stored in an .rds file containing meteo data as required
by the airGR family (see
\url{https://odelaigue.github.io/airGR/page_1_get_started.html)}. Required
are columns listed under Details
}
\details{
Required columns:
\itemize{
\item DatesR: dates in the POSIXt or Date format
\item P: average precipitation [mm/day]
\item T: catchment average air temperature [<U+2103>]
\item E: catchment average potential evapotranspiration [mm/day]
\item Qmm: outlet discharge [mm/day]
}
}
\examples{
load_meteo_data("D:/input/airGR/HSU_2044.rds")
}
