% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{merge_snow_runoff_sim}
\alias{merge_snow_runoff_sim}
\title{Merge Snow and Runoff Simulation Results}
\usage{
merge_snow_runoff_sim(simulation_results, snow_module_results)
}
\arguments{
\item{simulation_results}{A list returned by \code{\link{simulate_model}}, containing
simulated runoff and related metadata.}

\item{snow_module_results}{A list returned by \code{\link{simulate_snow}}, containing
snow-related outputs such as SWE, melt, and precipitation components.}
}
\value{
A list identical to \code{simulation_results}, but with additional elements:
\itemize{
\item \code{SWE}: Snow water equivalent.
\item \code{psolid}: Solid precipitation.
\item \code{pliquid}: Liquid precipitation.
\item \code{melt}: Meltwater.
\item \code{surface_water_runoff}: Combined snow runoff.
\item \code{more_info$snow_module}: Snow model output metadata.
}
}
\description{
Combines the output of a snow module simulation with the results of a hydrological
runoff simulation. Adds snow-related variables and snow model metadata to the
runoff simulation results.
}
\examples{
\dontrun{
merged_results <- merge_snow_runoff_sim(simulation_results, snow_module_results)
}
}
\seealso{
\code{\link{simulate_model}}, \code{\link{simulate_snow}}
}
