% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{optim_fn}
\alias{optim_fn}
\title{Function to be optimized during calibration}
\usage{
optim_fn(
  ParamOptim,
  hydro_data,
  split_indices,
  model,
  input,
  error_crit,
  cal_maximize,
  cal_q_transfo,
  do_transfo_param
)
}
\arguments{
\item{ParamOptim}{vector of model parameters to be optimized}

\item{hydro_data}{a list or data frame containing observed runoff loaded
with \code{\link{load_meteo_data}}}

\item{split_indices}{indices indicating the
warm-up and calibration period, i.e. the output of
\code{\link{split_data_set}}}

\item{model}{a string indicating the hydrological model}

\item{input}{model input returned by \code{\link{create_input}}}

\item{error_crit}{a string of a function indicating an error criterion, e.g.
\code{\link[hydroGOF]{KGE}}, or others from the \code{\link{hydroGOF}}
which are usable by \code{\link{calc_hydroGOF}}
functions}

\item{cal_maximize}{a logical indicating if function should be
maximized}

\item{cal_q_transfo}{a string indicating how runoff should be transformed
with \code{\link{transfo_q}}}

\item{do_transfo_param}{logical indicating
whether calibration is done with real or transformed parameters}
}
\value{
error criterion to be optimized
}
\description{
Function that takes global parameters (see Details) and performs
optimization to find best parameter set
}
\details{
Simulates a hydrological model during warm-up and calibration
period, and calculates an error criterion \code{error_crit} on the
calibration period only.

Replace NA in error criterion, which is generated when simulated runoff is
\itemize{
\item solely NAs: set a particularly bad value, i.e. +/- 1e10
(dependent on minimization or maximization)
\item solely 0s for \code{\link[hydroGOF]{KGE}}: set to asymptotic value,
i.e. 1 - sqrt(3)
\item for all other cases it throws an error
}
}
\note{
: Qsim is converted to numeric to similar to Qobs, this is important for
TUW for example, which returns a matrix 1 x time, for the lumped case.
todo:
\enumerate{
\item include a spatial explicit version
\item exclude warning for other NA cases
in error criterion as described above
\item Consider to link KGE etc with cal_maximize
}
}
\examples{
# differential evolution
cal_results <- DEoptim::DEoptim(
  fn = optim_fn, hydro_data = hydro_data, split_indices = split_indices,
  model = model, input = input, error_crit = error_crit, cal_maximize = cal_maximize,
  cal_q_transfo = cal_q_transfo, do_transfo_param = do_transfo_param,
  lower = lower, upper = upper,
  control = DEoptim::DEoptim.control(
    NP = cal_par[[model]][["DEoptim"]][["NP"]],
    itermax = cal_par[[model]][["DEoptim"]][["itermax"]],
    trace = 10
  )
)
}
\seealso{
\code{\link{calibrate_model}}, \code{\link{calibrate_model}},
\code{\link{create_input}}, \code{\link{load_meteo_data}},
\code{\link{calc_hydroGOF}}
}
