% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{split_data_set}
\alias{split_data_set}
\title{split a date vector}
\usage{
split_data_set(date, start_end_date_vec)
}
\arguments{
\item{date}{a vector of dates (e.g. Dates, Posix)}

\item{start_end_date_vec}{vector of length six, with entries start and end
date for warm-up, start and end date for calibration and start and end date
for validation, in this order.}
}
\value{
a list with three vectors containing indices \code{ind_warm},
\code{ind_cal}, \code{ind_val}
}
\description{
Splits a date vector hardcoded into three time periods, i.e. warm up,
calibration and validation period
}
\examples{
split_data_set(
  seq(from = as.Date("1981-01-01"), to = as.Date("2020-12-31"), by = "days"),
  c("1981-01-01", "1982-12-31", "1983-01-01", "2000-12-31", "2001-01-01", "2020-12-31")
)
}
