% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{subset_simulations}
\alias{subset_simulations}
\title{Subset Simulation Results by Index}
\usage{
subset_simulations(ind, simulation_results, keep_more_info = FALSE)
}
\arguments{
\item{ind}{A vector of indices to subset the simulation results.}

\item{simulation_results}{A list containing simulation results, where each element (except \code{"more_info"}) is assumed to be indexable.}

\item{keep_more_info}{Logical; if \code{TRUE}, the \code{"more_info"} field is retained in the output. Default is \code{FALSE}.}
}
\value{
A list of simulation results subset to the specified indices. If \code{keep_more_info} is \code{TRUE}, the \code{"more_info"} field is included unchanged.
}
\description{
This function subsets the elements of a simulation results list to the specified indices.
It excludes the \code{"more_info"} field by default but can optionally retain it.
}
\details{
This function assumes that all elements of \code{simulation_results}, except \code{"more_info"}, can be subset using the provided indices.
A future enhancement could include a check to ensure that each field is indeed subsettable.
}
\examples{
\dontrun{
results <- list(
  data = list(1:10, 11:20, 21:30),
  metrics = list(a = 1:3, b = 4:6),
  more_info = list(description = "Full simulation run")
)
subset_simulations(1:2, results, keep_more_info = TRUE)
}

}
