% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{transfo_param}
\alias{transfo_param}
\title{Parameter transformation}
\usage{
transfo_param(param, direction, model, cal_parameter = default_cal_par)
}
\arguments{
\item{param}{vector of model parameters}

\item{direction}{string indicating from real to transformed \code{"RT"} or
vice-versa \code{"TR"}.}

\item{model}{a string specifying the hydrological model}
}
\value{
a vector with transformed parameters
}
\description{
Transforms model parameters to hypercube and vice versa
}
\details{
For airGR models the functions \code{\link[airGR]{TransfoParam}} of package
\code{\link{airGR}}  are applied.
Model combinations as \code{"CemaNeigeGR4J"} are allowed.

For other models the parameter space is transformed to [0,1] with
\code{\link{norm_minmax}}
}
\note{
todo: 1) CemaNeigeHyst is not implemented yet
2) how do the links work without airGR installed?
3) implement functionality that airGR functions are not used at all
}
\examples{
# scale a parameter set for model "TUW" to [0,1] and back
param <- c(1, 2, 3, -1, 1, 1, 200, 10, 1, 15, 100, 50, 2, 15, 50)
scaled <- transfo_param(param, "RT", "TUW")
rescaled <- transfo_param(scaled, "TR", "TUW")
# scale a parameter set for "CemaNeigeGR4J"  to [-9.99,9.99] and back
param <- c(1000, 2, 250, 7, .2, 109.0365)
scaled <- transfo_param(param, "RT", "CemaNeigeGR4J")
rescaled <- transfo_param(scaled, "TR", "CemaNeigeGR4J")
}
\seealso{
\code{\link[airGR]{TransfoParam}}
}
