% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{transfo_q}
\alias{transfo_q}
\title{Transforms runoff}
\usage{
transfo_q(Q, q_transfo_type = "none")
}
\arguments{
\item{Q}{vector, matrix, data.frame etc of runoff values}

\item{q_transfo_type}{a string indicating how the runoff should be
transformed. Currently \code{"none"}, \code{"sqrt"}, \code{"inv"} and
\code{"log"} are supported.}
}
\value{
transformed runoff in same format as input
}
\description{
Inverse, sqrt and no transform of runoff, log is not advised for KGE, see
airGR or Santos 2018
}
\note{
todo: discuss what to do with Infinity
}
\examples{
transfo_q(array(0:10, c(2, 5)), "log")
}
