% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{validate_model}
\alias{validate_model}
\title{Validate model}
\usage{
validate_model(Qsim, Qobs, val_crit_transfo = "KGE__none")
}
\arguments{
\item{Qsim}{vector with simulated runoff}

\item{Qobs}{vector with observed runoff}

\item{val_crit_transfo}{a vector of strings specifying validation criteria and a
runoff transformation separated by a \code{"__"}. Supported are validation
criteria from the \code{\link{hydroGOF}} package usable by the
\code{\link{calc_hydroGOF}} function, for supported runoff transformations
please refer to \code{\link{transfo_q}}}
}
\value{
a long data frame with columns \code{crit} indicating the used
validation criterion, \code{transfo} for the used runoff transformation and
\code{value}.
}
\description{
Calculates validation measures for different transformation types
}
\examples{
validate_model(
  1:10, seq(2, 11),
  c("KGE__log", "NSE__inv", "VE__none", "pbias__none")
)
}
