% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{write_ascii}
\alias{write_ascii}
\title{Writes ascii results}
\usage{
write_ascii(
  file,
  calibration_results,
  validation_results,
  equally_spaced = TRUE
)
}
\arguments{
\item{file}{filename}

\item{calibration_results}{a list containing the calibration results from
\code{\link{calibrate_model}}. Only a vector with calibrated model
parameters are written out}

\item{validation_results}{a data frame containing the validation results from
\code{\link{validate_model}}.}

\item{equally_spaced}{(optional) a logical indicating if a equally spaced
output should be written out}
}
\value{
a logical if the file was successfully written
}
\description{
Writes an ascii overview of parameters and validation results
}
\examples{
write_ascii("results.txt", calibration_results, validation_results)
}
