% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{calc_subseasonal_validation_results}
\alias{calc_subseasonal_validation_results}
\title{Calculate Subseasonal Validation Metrics}
\usage{
calc_subseasonal_validation_results(
  val_subseason,
  dates,
  ind,
  period_name,
  col_name = "period",
  Qsim,
  Qobs,
  val_crit_transfo = "KGE__none"
)
}
\arguments{
\item{val_subseason}{A named list where each element is a character vector of two-digit month codes
(e.g., \code{"06"}, \code{"07"}) defining the months for each subseasonal period.}

\item{dates}{A vector of dates (e.g., of class \code{Date} or \code{POSIXct}) corresponding to the time series.}

\item{ind}{A vector of indices used to subset the hydrological data.}

\item{period_name}{A string indicating the name of the period (e.g., \code{"calibration"} or \code{"validation"}),
which will be added as a column in the output.}

\item{col_name}{A string specifying the name of the additional column to be added to the output
(default is \code{"period"}).}

\item{Qsim}{A numeric vector of simulated runoff values.}

\item{Qobs}{A numeric vector of observed runoff values.}

\item{val_crit_transfo}{A character vector specifying validation criteria and runoff transformations,
separated by \code{"__"}. Optionally, a third part can specify a lambda parameter. Supported criteria
are those from the \code{\link[hydroGOF]{hydroGOF}} package and must be compatible with
\code{\link{calc_hydroGOF}}. Supported transformations are described in \code{\link{transfo_q}}.}
}
\value{
A data frame similar to the output of \code{\link{validate_model}}, but with two additional
columns: one for the subseasonal period (e.g., \code{"spring"}, \code{"summer"}) and one for the period name
(e.g., \code{"calibration"}), as specified by \code{col_name} and \code{period_name}.
}
\description{
Computes validation metrics for specified subseasonal periods within a hydrological dataset.
For each named period in \code{val_subseason}, the function subsets the data using the provided indices
and calculates performance metrics using \code{\link{calc_validation_results}}.
}
\examples{
perf_cal <- calc_subseasonal_validation_results(
  val_subseason = list(
    spring = c("02", "03", "04", "05"),
    summer = c("06", "07", "08", "09")
  ),
  dates = hydro_data$BasinObs$DatesR,
  ind = split_indices$ind_cal,
  period_name = "calibration",
  col_name = "period",
  Qsim = simulation_results$Qsim,
  Qobs = Qobs,
  val_crit_transfo = c(
    "KGE__none", "NSE__none", "VE__none", "pbias__none",
    "KGE__inv", "NSE__inv",
    "KGE__sqrt", "NSE__sqrt"
  )
)

}
\seealso{
\code{\link{calc_validation_results}}, \code{\link{validate_model}},
\code{\link{calc_hydroGOF}}, \code{\link{transfo_q}}
}
