% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{calc_validation_results}
\alias{calc_validation_results}
\title{Calculates validation results for a subset period}
\usage{
calc_validation_results(
  ind,
  period_name,
  col_name = "period",
  Qsim,
  Qobs,
  val_crit_transfo = "KGE_none"
)
}
\arguments{
\item{ind}{indices used for subsetting \code{Qsim} and \code{Qobs}}

\item{period_name}{entries in column \code{col_name} naming the subset period}

\item{col_name}{additional column name in returned data frame (default is
period)}

\item{Qsim}{vector with simulated runoff}

\item{Qobs}{vector with observed runoff}

\item{val_crit_transfo}{a vector of strings specifying validation criteria and a
runoff transformation separated by a \code{"_"}. supported are validation
criteria from the \code{\link{hydroGOF}} package usable by the
\code{\link{calc_hydroGOF}} function, for supported runoff transformations
please refer to \code{\link{transfo_q}}}
}
\value{
a data frame as \code{\link{validate_model}} but with an additional
columns naming the period
}
\description{
Helper function which calls \code{\link{validate_model}} and subsets \code{Qsim}
and \code{Qobs} with \code{ind}, and returns a new data frame with an
additional column called \code{col_name} with entries \code{period_name}
}
\examples{
calc_validation_results(1:5, "summer", "season", 1:10, seq(0, 9))
}
\seealso{
\code{\link{validate_model}}, \code{\link{calc_hydroGOF}},
\code{\link{transfo_q}}
}
