% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{find_monthly_indices}
\alias{find_monthly_indices}
\title{Monthly indices}
\usage{
find_monthly_indices(date, months, ind = seq_along(date))
}
\arguments{
\item{date}{a vector of dates (e.g. Dates, Posix)}

\item{months}{a vector of months in double digits strings, e.g.
\code{c("02",  "03")}}

\item{ind}{a vector of indices which can be used to subset the time dependent
elements in input.}
}
\value{
a vector with indices
}
\description{
Returns monthly indices for specified months within a date vector
}
\examples{
find_monthly_indices(seq(from = as.Date("1981-01-01"), to = as.Date("2020-12-31"), by = "days"), c("02", "03"))
}
