% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{norm_minmax}
\alias{norm_minmax}
\title{Min-max normalization and vice versa}
\usage{
norm_minmax(x, min, max, direction = "RT")
}
\arguments{
\item{x}{data to be scaled, can be a vector, array or matrix}

\item{min}{lower bound}

\item{max}{upper bound}

\item{direction}{string indicating from real to transformed \code{"RT"} or
vice-versa \code{"TR"}. For direction \code{"TR"} the lower and upper
bounds are required inputs.}
}
\value{
transformed or re-transformed data of same format as \code{x}
}
\description{
Scales data to [0,1] with min and max of applied data or with specified
bounds
}
\note{
todo: \code{function(x, min = min(x), max = max(x), direction = "RT")}
does not work, other solutions?
default values for min and max only valid if
direction is also RT, how can this be required
}
\examples{
vec <- runif(10, -5, 10)
norm_minmax(vec, min(vec), max(vec))
vecT <- norm_minmax(vec, -5, 10)
# re-scale to vec
norm_minmax(vecT, -5, 10, "TR")
}
