% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{save_airGR_plot}
\alias{save_airGR_plot}
\title{save an airGR overview plot to pdf}
\usage{
save_airGR_plot(file, simulation_results, Qobs)
}
\arguments{
\item{file}{the filename of the saved plot}

\item{simulation_results}{output from \code{\link{simulate_model}}, i.e. the
output object of the \code{\link{airGR}} function
\code{\link[airGR]{RunModel}}}

\item{Qobs}{a vector of observed runoff}
}
\value{
a logical if file was written successfully
}
\description{
use the \code{\link{airGR}} \code{\link[airGR]{plot}} function
}
\examples{
save_airGR_plot("airGR_plot.pdf", simulation_results, Qobs)
}
