% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{save_cal_val_plot}
\alias{save_cal_val_plot}
\title{Save a calibration validation plot}
\usage{
save_cal_val_plot(file, BasinObs, Qsim, split_indices)
}
\arguments{
\item{file}{the filename of the saved plot}

\item{BasinObs}{data frame with time series of input data from
\code{\link{load_meteo_data}}}

\item{Qsim}{vector with simulated runoff}

\item{split_indices}{a list of indices from \code{\link{split_data_set}}
containing the elements \code{ind_cal} and \code{ind_val}}
}
\value{
a logical if the file was successfully written
}
\description{
Create and save a plot containing over the calibration and validation period
}
\details{
Plots observed and simulated runoff and displays certain error metrics
}
\note{
todo: include a seasonal rolling mean
}
\examples{
save_cal_val_plot("cal_val.pdf", BasinObs, split_indices)
}
