% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{set_airGR_par}
\alias{set_airGR_par}
\title{Create default calibration settings for the airGR family}
\usage{
set_airGR_par(model)
}
\arguments{
\item{model}{a model string}
}
\value{
a list of calibration settings
}
\description{
Create default calibration settings for the airGR family
}
\examples{
set_airGR_par("GR4J")
}
