% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_and_settings.R
\name{transfo_q}
\alias{transfo_q}
\title{Apply transformation to runoff data}
\usage{
transfo_q(Q, q_transfo_type = "none", lambda = 0.25, ...)
}
\arguments{
\item{Q}{A numeric vector, matrix, or data frame of runoff values.}

\item{q_transfo_type}{A string specifying the transformation type. Options are:
"none", "sqrt", "inv", "log", "boxcox", "boxcoxsantos", and "power".}

\item{lambda}{A numeric value used for Box-Cox and power transformations. Default is 0.25.}

\item{...}{Additional arguments passed to mean(), e.g., na.rm = TRUE, used in boxcoxsantos.}
}
\value{
Transformed runoff data in the same format as input.
}
\description{
Supports inverse, square root, power, and Box-Cox transformations of runoff data.
Log transformation is included but generally not recommended for KGE calculations
(see airGR or Santos, 2018). Use Use \code{\link{KGEtang}} instead.
}
\note{
Consider how to handle infinite values resulting from transformations.
}
\examples{
transfo_q(array(0:10, c(2, 5)), "sqrt")
}
